Reduce code duplication between modules
