+++ native/include/domain.h
@@ -49,12 +49,13 @@ typedef struct mem mem_t;
 /* status of the domain as read/store in httpd. */
 struct domaininfo
 {
+    /* NOTE: Due to `loc_get_id`, struct MUST begin with id */
+    int id;                    /* id in table */
     char domain[DOMAINNDSZ];   /* domain value */
     char JVMRoute[JVMROUTESZ]; /* corresponding node */
     char balancer[BALANCERSZ]; /* name of the balancer */
 
     apr_time_t updatetime; /* time of last received message */
-    int id;                /* id in table */
 };
 typedef struct domaininfo domaininfo_t;
