+++ native/mod_manager/CMakeLists.txt
@@ -16,6 +16,7 @@ SET(MOD_MANAGER_SRCS    ${PROJECT_SOURCE_DIR}/mod_manager.c
         ${PROJECT_SOURCE_DIR}/host.c
         ${PROJECT_SOURCE_DIR}/node.c
         ${PROJECT_SOURCE_DIR}/sessionid.c
+        ${PROJECT_SOURCE_DIR}/../common/common.c
 )
 
 INCLUDE_DIRECTORIES("${PROJECT_BINARY_DIR}")
@@ -24,4 +25,9 @@ INCLUDE_DIRECTORIES("${PROJECT_INCLUDE_DIR}")
 ADD_LIBRARY(mod_manager MODULE ${MOD_MANAGER_SRCS})
 SET_TARGET_PROPERTIES(mod_manager PROPERTIES PREFIX "")
 SET_TARGET_PROPERTIES(mod_manager PROPERTIES SUFFIX ".so")
-TARGET_LINK_LIBRARIES(mod_manager ${APR_LIBRARIES} ${APRUTIL_LIBRARIES} ${APACHE_LIBRARY})
+
+IF(WIN32)
+    TARGET_LINK_LIBRARIES(mod_manager ${PROXY_LIBRARY} ${APR_LIBRARIES} ${APRUTIL_LIBRARIES} ${APACHE_LIBRARY})
+ELSE()
+    TARGET_LINK_LIBRARIES(mod_manager ${APR_LIBRARIES} ${APRUTIL_LIBRARIES} ${APACHE_LIBRARY})
+ENDIF()
