+++ native/mod_manager/context.c
@@ -44,6 +44,9 @@
 
 #include "mod_manager.h"
 
+#include "common.h"
+
+
 static mem_t *create_attach_mem_context(char *string, unsigned *num, int type, int create, apr_pool_t *p,
                                         slotmem_storage_method *storage)
 {
@@ -181,16 +184,6 @@ apr_status_t remove_context(mem_t *s, int id)
     return s->storage->release(s->slotmem, id);
 }
 
-static apr_status_t loc_get_id(void *mem, void *data, apr_pool_t *pool)
-{
-    struct counter *count = (struct counter *)data;
-    contextinfo_t *ou = (contextinfo_t *)mem;
-    (void)pool;
-    *count->values = ou->id;
-    count->values++;
-    count->count++;
-    return APR_SUCCESS;
-}
 
 /*
  * get the ids for the used (not free) contexts in the table
